# packages/openapi-client/openapi.json
{"openapi":"3.0.0","paths":{"/v1/health":{"get":{"operationId":"HealthController_getHealth_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/v1/public/pages/{slug}":{"get":{"operationId":"PublicPagesController_getPublishedPage_v1","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}},{"name":"locale","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["public"]}},"/v1/admin/pages":{"get":{"operationId":"AdminPagesController_listPages_v1","parameters":[],"responses":{"200":{"description":""}},"tags":["admin/pages"]},"post":{"operationId":"AdminPagesController_create_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"page":{"id":"…","slug":"about","locale":"uk","status":"draft"}}}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}/preview-token":{"post":{"operationId":"AdminPagesController_createPreviewToken_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"token":"jwt...","expiresAt":"2026-01-01T00:00:00.000Z"}}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}":{"get":{"operationId":"AdminPagesController_getPage_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminPageResponseDto"}}}}},"tags":["admin/pages"]},"patch":{"operationId":"AdminPagesController_update_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"page":{"id":"…","slug":"about-us"}}}}}}},"tags":["admin/pages"]},"delete":{"operationId":"AdminPagesController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true}}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}/publish":{"post":{"operationId":"AdminPagesController_publish_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true}}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}/unpublish":{"post":{"operationId":"AdminPagesController_unpublish_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true}}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}/duplicate":{"post":{"operationId":"AdminPagesController_duplicate_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminPageShortResponseDto"}}}}},"tags":["admin/pages"]}},"/v1/admin/pages/{id}/blocks":{"post":{"operationId":"AdminPagesController_createBlock_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBlockDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"block":{"id":"…","type":"hero","variant":"v1","order":8}}}}}}},"tags":["admin/pages"]}},"/v1/public/preview/pages/{id}":{"get":{"operationId":"PublicPreviewController_getPreviewPage_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["public/preview"]}},"/v1/admin/blocks/{id}":{"patch":{"operationId":"AdminBlocksController_updateBlock_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBlockDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true}}}}}},"tags":["admin/blocks"]},"delete":{"operationId":"AdminBlocksController_remove_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true}}}}}},"tags":["admin/blocks"]}},"/v1/admin/blocks/{id}/move":{"post":{"operationId":"AdminBlocksController_move_v1","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MoveBlockDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"ok":true,"moved":true}}}}}},"tags":["admin/blocks"]}},"/v1/auth/login":{"post":{"operationId":"AuthController_login_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/v1/auth/me":{"get":{"operationId":"AuthController_me_v1","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["auth"]}},"/v1/auth/logout":{"post":{"operationId":"AuthController_logout_v1","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"Landing → Admin → CRM API","description":"Backend API for landing/admin and future CRM modules","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AdminBlockDto":{"type":"object","properties":{"id":{"type":"string"},"type":{"type":"string"},"variant":{"type":"string"},"order":{"type":"number"},"data":{"type":"object","additionalProperties":true,"nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","type","variant","order","data","createdAt","updatedAt"]},"AdminPageDto":{"type":"object","properties":{"id":{"type":"string"},"slug":{"type":"string"},"locale":{"type":"string"},"status":{"type":"string","enum":["draft","published"]},"seo":{"type":"object","additionalProperties":true,"nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"publishedAt":{"type":"string","format":"date-time","nullable":true},"blocks":{"type":"array","items":{"$ref":"#/components/schemas/AdminBlockDto"}}},"required":["id","slug","locale","status","seo","createdAt","updatedAt","publishedAt","blocks"]},"AdminPageResponseDto":{"type":"object","properties":{"page":{"$ref":"#/components/schemas/AdminPageDto"}},"required":["page"]},"CreatePageDto":{"type":"object","properties":{"slug":{"type":"string","example":"about"},"locale":{"type":"string","enum":["uk","en"],"example":"uk"}},"required":["slug","locale"]},"AdminPageShortDto":{"type":"object","properties":{"id":{"type":"string"},"slug":{"type":"string"},"locale":{"type":"string"},"status":{"type":"string","enum":["draft","published"]}},"required":["id","slug","locale","status"]},"AdminPageShortResponseDto":{"type":"object","properties":{"page":{"$ref":"#/components/schemas/AdminPageShortDto"}},"required":["page"]},"UpdatePageDto":{"type":"object","properties":{"slug":{"type":"string","example":"about-us"}}},"CreateBlockDto":{"type":"object","properties":{"type":{"type":"string","example":"hero"},"variant":{"type":"string","example":"v1"},"data":{"type":"object","additionalProperties":true}},"required":["type","variant"]},"UpdateBlockDto":{"type":"object","properties":{"data":{"type":"object","additionalProperties":true}},"required":["data"]},"MoveBlockDto":{"type":"object","properties":{"direction":{"type":"string","enum":["up","down"]}},"required":["direction"]},"LoginDto":{"type":"object","properties":{}}}}}